+++ base/src/index.ts
@@ -17,6 +17,10 @@ export type DisposeFunction = () => void
 
 export type StreamSubscribe = <T>(stream: any, observer: Observer<T>) => DisposeFunction | void
 
+export interface DevToolEnabledSource {
+  _isCycleSource: string;
+}
+
 export interface StreamAdapter {
   adapt: <T>(originStream: any, originStreamSubscribe: StreamSubscribe) => any;
   remember: <T>(stream: any) => any;
@@ -101,6 +105,9 @@ function callDrivers(drivers: DriversDefinition,
         );
       } else {
         sources[name] = driverOutput;
+      }
+      if (sources[name] && typeof sources[name] === 'object') {
+        (<DevToolEnabledSource> sources[name])._isCycleSource = name;
       }
     }
   }