+++ dom/src/HTMLSource.ts
@@ -1,5 +1,5 @@
 import xs, {Stream} from 'xstream';
-import {StreamAdapter} from '@cycle/base';
+import {StreamAdapter, DevToolEnabledSource} from '@cycle/base';
 import {DOMSource, EventsFnOptions} from './DOMSource';
 import xsSA from '@cycle/xstream-adapter';
 
@@ -8,20 +8,28 @@ export class HTMLSource implements DOMSource {
   private _empty$: any;
 
   constructor(html$: Stream<string>,
-              private runSA: StreamAdapter) {
+              private runSA: StreamAdapter,
+              private _name: string) {
     this._html$ = html$;
     this._empty$ = runSA.adapt(xs.empty(), xsSA.streamSubscribe);
   }
 
-  elements(): any {
-    return this.runSA.adapt(this._html$, xsSA.streamSubscribe);
+  public elements(): any {
+    const out: DevToolEnabledSource = this.runSA.adapt(
+      this._html$,
+      xsSA.streamSubscribe
+    );
+    out._isCycleSource = this._name;
+    return out;
   }
 
   public select(selector: string): DOMSource {
-    return new HTMLSource(xs.empty(), this.runSA);
+    return new HTMLSource(xs.empty(), this.runSA, this._name);
   }
 
   public events(eventType: string, options?: EventsFnOptions): any {
-    return this._empty$;
+    const out: DevToolEnabledSource = this._empty$;
+    out._isCycleSource = this._name;
+    return out;
   }
 }