+++ http/src/http-driver.ts
@@ -128,14 +128,17 @@ function normalizeRequestInput(reqOptions: RequestInput): RequestOptions {
 function makeRequestInputToResponse$(runStreamAdapter: StreamAdapter) {
   return function requestInputToResponse$(reqInput: RequestInput): MemoryStream<Response> & ResponseStream {
     let response$ = createResponse$(reqInput).remember();
-    /* tslint:disable:no-empty */
-    response$.addListener({next: () => {}, error: () => {}, complete: () => {}});
-    /* tslint:enable:no-empty */
+    let reqOptions = softNormalizeRequestInput(reqInput);
+    if (!reqOptions.lazy) {
+      /* tslint:disable:no-empty */
+      response$.addListener({next: () => {}, error: () => {}, complete: () => {}});
+      /* tslint:enable:no-empty */
+    }
     response$ = (runStreamAdapter) ?
       runStreamAdapter.adapt(response$, XStreamAdapter.streamSubscribe) :
       response$;
     Object.defineProperty(response$, 'request', <PropertyDescriptor> {
-      value: softNormalizeRequestInput(reqInput),
+      value: reqOptions,
       writable: false,
     });
     return <MemoryStream<Response> & ResponseStream> response$;
@@ -143,11 +146,11 @@ function makeRequestInputToResponse$(runStreamAdapter: StreamAdapter) {
 }
 
 export function makeHTTPDriver(): Function {
-  function httpDriver(request$: Stream<RequestInput>, runSA: StreamAdapter): HTTPSource {
+  function httpDriver(request$: Stream<RequestInput>, runSA: StreamAdapter, name: string): HTTPSource {
     let response$$ = request$
       .map(makeRequestInputToResponse$(runSA))
       .remember();
-    let httpSource = new MainHTTPSource(response$$, runSA, []);
+    let httpSource = new MainHTTPSource(response$$, runSA, name, []);
     /* tslint:disable:no-empty */
     response$$.addListener({next: () => {}, error: () => {}, complete: () => {}});
     /* tslint:enable:no-empty */