+++ base/CHANGELOG.md
@@ -1,3 +1,23 @@
+<a name="4.1.1"></a>
+## 4.1.1 (2016-08-21)
+
+
+### Bug Fixes
+
+* **base:** support noop driver with undefined source ([02fb507](https://github.com/cyclejs/cyclejs/tree/master/packages/base/commit/02fb507))
+
+
+
+<a name="4.1.0"></a>
+# 4.1.0 (2016-08-20)
+
+
+### Features
+
+* **base:** add devtool flag to each source stream ([e6eff9e](https://github.com/cyclejs/cyclejs/tree/master/packages/base/commit/e6eff9e))
+
+
+
 <a name="4.0.2"></a>
 ## 4.0.2 (2016-07-31)