+++ most-adapter/CHANGELOG.md
@@ -1,3 +1,28 @@
+<a name="4.0.1"></a>
+## 4.0.1 (2016-08-11)
+
+
+### Bug Fixes
+
+* **most-adapter:** make stream subscription lazy, implement proper disposal ([242bfa5](https://github.com/cyclejs/cyclejs/tree/master/packages/most-adapter/commit/242bfa5))
+
+
+
+<a name="4.0.0"></a>
+# 4.0.0 (2016-08-09)
+
+
+### Features
+
+* **most-adapter:** update most to v1.0.0 ([e2b2267](https://github.com/cyclejs/cyclejs/tree/master/packages/most-adapter/commit/e2b2267))
+
+
+### BREAKING CHANGES
+
+* most-adapter: Updates most to v1.0.0, which is known to be a breaking version in relation to the previous v0.19
+
+
+
 <a name="3.0.0"></a>
 # 3.0.0 (2016-07-16)