+++ most-run/CHANGELOG.md
@@ -1,3 +1,28 @@
+<a name="4.1.0"></a>
+# 4.1.0 (2016-08-21)
+
+
+### Features
+
+* **most-run:** update cycle base run() to v4.1 ([e7801fa](https://github.com/cyclejs/cyclejs/tree/master/packages/most-run/commit/e7801fa))
+
+
+
+<a name="4.0.0"></a>
+# 4.0.0 (2016-08-09)
+
+
+### Features
+
+* **most-run:** update most to v1.0.0 ([ce35480](https://github.com/cyclejs/cyclejs/tree/master/packages/most-run/commit/ce35480))
+
+
+### BREAKING CHANGES
+
+* most-run: Updates most to v1.0.0 which is known to be a breaking version in relation to the previous v0.19
+
+
+
 <a name="3.0.1"></a>
 ## 3.0.1 (2016-07-16)