+++ rxjs-adapter/CHANGELOG.md
@@ -1,3 +1,13 @@
+<a name="3.0.2"></a>
+## 3.0.2 (2016-08-09)
+
+
+### Bug Fixes
+
+* **rxjs-adapter:** make most.js streams fail isValidStream test ([fde59c4](https://github.com/cyclejs/cyclejs/tree/master/packages/rxjs-adapter/commit/fde59c4))
+
+
+
 <a name="3.0.1"></a>
 ## 3.0.1 (2016-07-16)